requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1

# Database
aiosqlite==0.19.0

# Web scraping and monitoring for tax updates
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
aiohttp==3.9.1

# Scheduling for automated checks
apscheduler==3.10.4

# Logging and error tracking
loguru==0.7.2
sentry-sdk==1.38.0

# Data validation
pydantic==2.5.2

# Date handling
python-dateutil==2.8.2

# Internationalization
babel==2.13.1

# Configuration
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
